+++ compiler/add_clause.m
@@ -480,9 +480,9 @@ select_applicable_modes(ModuleInfo, MaybeAnnotatedArgTerms, VarSet,
                             color_as_incorrect([words("no need to restrict"),
                                 words("a clause for it to that mode.")]) ++
                             [nl],
-                        % XXX SEVERITY
-                        Spec = spec($pred, severity_warning(warn_simple_code),
-                            phase_pt2h, Context, Pieces),
+                        Severity = severity_warning(warn_redundant_code),
+                        Spec = spec($pred, Severity, phase_pt2h,
+                            Context, Pieces),
                         !:Specs = [Spec | !.Specs]
                     ;
                         Warn = no