+++ compiler/convert_parse_tree.m
@@ -1976,8 +1976,8 @@ report_int_imp_fim(IntFIMSpecMap, FIMSpec, !ImpFIMSpecMap, !Specs) :-
         words("pragma in the interface section is here."), nl],
     ImpMsg = msg(ImpContext, ImpPieces),
     IntMsg = msg(IntContext, IntPieces),
-    Spec = error_spec($pred, severity_warning(warn_simple_code),
-        phase_pt2h, [ImpMsg, IntMsg]),
+    Severity = severity_warning(warn_redundant_code),
+    Spec = error_spec($pred, Severity, phase_pt2h, [ImpMsg, IntMsg]),
     !:Specs = [Spec | !.Specs].
 
 %---------------------------------------------------------------------------%
@@ -2145,8 +2145,8 @@ classify_foreign_import_module(ItemFIM, !FIMSpecMap, !Specs) :-
             words("pragma for the same module/language combination"),
             words("is here."), nl],
         PrevMsg = msg(PrevContext, PrevPieces),
-        Spec = error_spec($pred, severity_warning(warn_simple_code),
-            phase_pt2h, [MainMsg, PrevMsg]),
+        Severity = severity_warning(warn_redundant_code),
+        Spec = error_spec($pred, Severity, phase_pt2h, [MainMsg, PrevMsg]),
         !:Specs = [Spec | !.Specs]
     else
         map.det_insert(FIMSpec, Context, !FIMSpecMap)