+++ compiler/simplify_goal.m
@@ -219,7 +219,7 @@ maybe_simplify_goal_to_true_or_fail(InstMap0, Goal0ContainsTrace,
         % - a call to `builtin.false/0'.
         Context = goal_info_get_context(GoalInfo0),
         ( if
-            simplify_do_warn_simple_code(!.Info),
+            simplify_do_warn_dodgy_simple_code(!.Info),
             not (
                 some [SubGoal] (
                     goal_contains_goal(Goal0, SubGoal),
@@ -238,7 +238,7 @@ maybe_simplify_goal_to_true_or_fail(InstMap0, Goal0ContainsTrace,
             Msg = simple_msg(Context,
                 [always(MainPieces),
                 verbose_only(verbose_always, VerbosePieces)]),
-            Spec = error_spec($pred, severity_warning(warn_simple_code),
+            Spec = error_spec($pred, severity_warning(warn_dodgy_simple_code),
                 phase_simplify(report_only_if_in_all_modes), [Msg]),
             simplify_info_add_message(Spec, !Info)
         else